Provide a NICKEL_NAME(...) shorthand
NICKEL_NAME(abc, abc) was rather common, so this makes that expressible
via NICKEL_NAME(abc).

Note that the macro varargs `...` must have an argument; it cannot
legally be left blank. We pass in an extra dummy argument to make this
truly legal (Clang will complain otherwise).

Also, MSVC's __VA_ARGS__ expansion works peculiarly, expanding as one
argument with a comma in it rather than as two. This is somehow fixed
via wrapping with an EXPAND(...) macro.