Warn about forgetting to invoke void function
If the user forgets the final `()` to invoke the function, they will now
receive a warning. This is only relevant for void-returning functions
where the result is not otherwise stored. For functions which return a
value, this is both impossible to check and unnecessary, because they
will probably receive other compile errors down the line.